create separate action + preset for Order Completed
