+++ packages/destination-actions/src/destinations/amplitude/event-schema.ts
@@ -1,4 +1,4 @@
-import { InputField } from '@segment/actions-core'
+import type { InputField } from '@segment/actions-core'
 
 /**
  * The common fields defined by Amplitude's events api
@@ -304,5 +304,77 @@ export const eventSchema: Record<string, InputField> = {
     default: {
       '@path': '$.context.library.name'
     }
+  },
+  use_batch_endpoint: {
+    label: 'Use Batch Endpoint',
+    description:
+      "If true, events are sent to Amplitude's `batch` endpoint rather than their `httpapi` events endpoint. Enabling this setting may help reduce 429s – or throttling errors – from Amplitude. More information about Amplitude's throttling is available in [their docs](https://developers.amplitude.com/docs/batch-event-upload-api#429s-in-depth).",
+    type: 'boolean',
+    default: false
+  },
+  userAgent: {
+    label: 'User Agent',
+    type: 'string',
+    description: 'The user agent of the device sending the event.',
+    default: {
+      '@path': '$.context.userAgent'
+    }
+  },
+  userAgentParsing: {
+    label: 'User Agent Parsing',
+    type: 'boolean',
+    description:
+      'Enabling this setting will set the Device manufacturer, Device Model and OS Name properties based on the user agent string provided in the userAgent field',
+    default: true
+  },
+  utm_properties: {
+    label: 'UTM Properties',
+    type: 'object',
+    description: 'UTM Tracking Properties',
+    properties: {
+      utm_source: {
+        label: 'UTM Source',
+        type: 'string'
+      },
+      utm_medium: {
+        label: 'UTM Medium',
+        type: 'string'
+      },
+      utm_campaign: {
+        label: 'UTM Campaign',
+        type: 'string'
+      },
+      utm_term: {
+        label: 'UTM Term',
+        type: 'string'
+      },
+      utm_content: {
+        label: 'UTM Content',
+        type: 'string'
+      }
+    },
+    default: {
+      utm_source: { '@path': '$.context.campaign.source' },
+      utm_medium: { '@path': '$.context.campaign.medium' },
+      utm_campaign: { '@path': '$.context.campaign.name' },
+      utm_term: { '@path': '$.context.campaign.term' },
+      utm_content: { '@path': '$.context.campaign.content' }
+    }
+  },
+  referrer: {
+    label: 'Referrer',
+    type: 'string',
+    description:
+      'The referrer of the web request. Sent to Amplitude as both last touch “referrer” and first touch “initial_referrer”',
+    default: {
+      '@path': '$.context.page.referrer'
+    }
+  },
+  min_id_length: {
+    label: 'Minimum ID Length',
+    description:
+      'Amplitude has a default minimum id lenght of 5 characters for user_id and device_id fields. This field allows the minimum to be overridden to allow shorter id lengths.',
+    allowNull: true,
+    type: 'integer'
   }
 }
