+++ packages/destination-actions/src/destinations/amplitude/logEvent/index.ts
@@ -18,161 +18,13 @@ export interface AmplitudeEvent extends Omit<Payload, 'products' | 'trackRevenue
   }
 }
 
-const revenueKeys = ['revenue', 'price', 'productId', 'quantity', 'revenueType']
-
-interface EventRevenue {
-  revenue?: number
-  price?: number
-  productId?: string
-  quantity?: number
-  revenueType?: string
-}
-
-function getRevenueProperties(payload: EventRevenue): EventRevenue {
-  if (typeof payload.revenue !== 'number') {
-    return {}
-  }
-
-  return {
-    revenue: payload.revenue,
-    revenueType: payload.revenueType ?? 'Purchase',
-    quantity: typeof payload.quantity === 'number' ? Math.round(payload.quantity) : undefined,
-    price: payload.price,
-    productId: payload.productId
-  }
-}
-
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Log Event',
   description: 'Send an event to Amplitude.',
   defaultSubscription: 'type = "track"',
-  fields: {
-    trackRevenuePerProduct: {
-      label: 'Track Revenue Per Product',
-      description:
-        'When enabled, track revenue with each product within the event. When disabled, track total revenue once for the event.',
-      type: 'boolean',
-      default: false
-    },
-    ...eventSchema,
-    products: {
-      label: 'Products',
-      description: 'The list of products purchased.',
-      type: 'object',
-      multiple: true,
-      properties: {
-        price: {
-          label: 'Price',
-          type: 'number',
-          description:
-            'The price of the item purchased. Required for revenue data if the revenue field is not sent. You can use negative values to indicate refunds.'
-        },
-        quantity: {
-          label: 'Quantity',
-          type: 'integer',
-          description: 'The quantity of the item purchased. Defaults to 1 if not specified.'
-        },
-        revenue: {
-          label: 'Revenue',
-          type: 'number',
-          description:
-            'Revenue = price * quantity. If you send all 3 fields of price, quantity, and revenue, then (price * quantity) will be used as the revenue value. You can use negative values to indicate refunds.'
-        },
-        productId: {
-          label: 'Product ID',
-          type: 'string',
-          description:
-            'An identifier for the item purchased. You must send a price and quantity or revenue with this field.'
-        },
-        revenueType: {
-          label: 'Revenue Type',
-          type: 'string',
-          description:
-            'The type of revenue for the item purchased. You must send a price and quantity or revenue with this field.'
-        }
-      },
-      default: {
-        '@path': '$.properties.products'
-      }
-    },
-    use_batch_endpoint: {
-      label: 'Use Batch Endpoint',
-      description:
-        "If true, events are sent to Amplitude's `batch` endpoint rather than their `httpapi` events endpoint. Enabling this setting may help reduce 429s – or throttling errors – from Amplitude. More information about Amplitude's throttling is available in [their docs](https://developers.amplitude.com/docs/batch-event-upload-api#429s-in-depth).",
-      type: 'boolean',
-      default: false
-    },
-    userAgent: {
-      label: 'User Agent',
-      type: 'string',
-      description: 'The user agent of the device sending the event.',
-      default: {
-        '@path': '$.context.userAgent'
-      }
-    },
-    userAgentParsing: {
-      label: 'User Agent Parsing',
-      type: 'boolean',
-      description:
-        'Enabling this setting will set the Device manufacturer, Device Model and OS Name properties based on the user agent string provided in the userAgent field',
-      default: true
-    },
-    utm_properties: {
-      label: 'UTM Properties',
-      type: 'object',
-      description: 'UTM Tracking Properties',
-      properties: {
-        utm_source: {
-          label: 'UTM Source',
-          type: 'string'
-        },
-        utm_medium: {
-          label: 'UTM Medium',
-          type: 'string'
-        },
-        utm_campaign: {
-          label: 'UTM Campaign',
-          type: 'string'
-        },
-        utm_term: {
-          label: 'UTM Term',
-          type: 'string'
-        },
-        utm_content: {
-          label: 'UTM Content',
-          type: 'string'
-        }
-      },
-      default: {
-        utm_source: { '@path': '$.context.campaign.source' },
-        utm_medium: { '@path': '$.context.campaign.medium' },
-        utm_campaign: { '@path': '$.context.campaign.name' },
-        utm_term: { '@path': '$.context.campaign.term' },
-        utm_content: { '@path': '$.context.campaign.content' }
-      }
-    },
-    referrer: {
-      label: 'Referrer',
-      type: 'string',
-      description:
-        'The referrer of the web request. Sent to Amplitude as both last touch “referrer” and first touch “initial_referrer”',
-      default: {
-        '@path': '$.context.page.referrer'
-      }
-    },
-    min_id_length: {
-      label: 'Minimum ID Length',
-      description:
-        'Amplitude has a default minimum id lenght of 5 characters for user_id and device_id fields. This field allows the minimum to be overridden to allow shorter id lengths.',
-      allowNull: true,
-      type: 'integer'
-    }
-  },
+  fields: eventSchema,
   perform: (request, { payload, settings }) => {
-    // Omit revenue properties initially because we will manually stitch those into events as prescribed
     const {
-      products = [],
-      trackRevenuePerProduct,
       time,
       session_id,
       userAgent,
@@ -182,7 +34,7 @@ const action: ActionDefinition<Settings, Payload> = {
       min_id_length,
       library,
       ...rest
-    } = omit(payload, revenueKeys)
+    } = payload
     const properties = rest as AmplitudeEvent
     let options
 
@@ -190,8 +42,8 @@ const action: ActionDefinition<Settings, Payload> = {
       properties.platform = properties.platform.replace(/ios/i, 'iOS').replace(/android/i, 'Android')
     }
 
-    if (library) {
-      if (library === 'analytics.js') properties.platform = 'Web'
+    if (library === 'analytics.js') {
+      properties.platform = 'Web'
     }
 
     if (time && dayjs.utc(time).isValid()) {
@@ -220,23 +72,9 @@ const action: ActionDefinition<Settings, Payload> = {
         ...(userAgentParsing && parseUserAgentProperties(userAgent)),
         // Make sure any top-level properties take precedence over user-agent properties
         ...removeUndefined(properties),
-        // Conditionally track revenue with main event
-        ...(products.length && trackRevenuePerProduct ? {} : getRevenueProperties(payload)),
         library: 'segment'
       }
     ]
-
-    for (const product of products) {
-      events.push({
-        ...properties,
-        // Or track revenue per product
-        ...(trackRevenuePerProduct ? getRevenueProperties(product as EventRevenue) : {}),
-        event_properties: product,
-        event_type: 'Product Purchased',
-        insert_id: properties.insert_id ? `${properties.insert_id}-${events.length + 1}` : undefined,
-        library: 'segment'
-      })
-    }
 
     const endpoint = payload.use_batch_endpoint
       ? 'https://api2.amplitude.com/batch'
