+++ packages/destination-actions/src/destinations/amplitude/orderCompleted/generated-types.ts
@@ -2,6 +2,35 @@
 
 export interface Payload {
   /**
+   * When enabled, track revenue with each product within the event. When disabled, track total revenue once for the event.
+   */
+  trackRevenuePerProduct: boolean
+  /**
+   * The list of products purchased.
+   */
+  products?: {
+    /**
+     * The price of the item purchased. Required for revenue data if the revenue field is not sent. You can use negative values to indicate refunds.
+     */
+    price?: number
+    /**
+     * The quantity of the item purchased. Defaults to 1 if not specified.
+     */
+    quantity?: number
+    /**
+     * Revenue = price * quantity. If you send all 3 fields of price, quantity, and revenue, then (price * quantity) will be used as the revenue value. You can use negative values to indicate refunds.
+     */
+    revenue?: number
+    /**
+     * An identifier for the item purchased. You must send a price and quantity or revenue with this field.
+     */
+    productId?: string
+    /**
+     * The type of revenue for the item purchased. You must send a price and quantity or revenue with this field.
+     */
+    revenueType?: string
+  }[]
+  /**
    * A readable ID specified by you. Must have a minimum length of 5 characters. Required unless device ID is present. **Note:** If you send a request with a user ID that is not in the Amplitude system yet, then the user tied to that ID will not be marked new until their first event.
    */
   user_id?: string | null
